lot of fixes and clarify the code in simplex

+++ Generation/Noise/simplex.h
@@ -1,6 +1,5 @@
 #ifndef SIMPLEX_H__
 #define SIMPLEX_H__
 #include "../Utils/utils.h"
-SDL_Surface* generate_simplex(int x, int y);
-
+SDL_Surface* generate_simplex(double x, double y);
 #endif
